Always remove space after star (if possible)
Because the current pattern is a non-greedy space match *at the end* of
the regex, it's never going to do anything (the regex will always match
with 0 spaces and stop there).

I assume the intent was to remove the (common) space separating the star
and text, removing the repetition should do that. Making the operator
greedy would break e.g. indented examples which probably isn't a good
idea.